backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
numpy==1.24.4; python_version < "3.10"
numpy==2.1.3; python_version >= "3.10"
faiss-cpu==1.9.0.post1; platform_system != "Windows"
rapidfuzz==3.9.7
orjson==3.10.12
pytest==8.3.3
